Added `odoo-helper test --recreate-db` option

+++ lib/test.bash
@@ -105,7 +105,8 @@ function test_module_impl {
 # test_get_or_create_db    # get db
 # test_get_or_create_db 1  # create new db
 function test_get_or_create_db {
-    local create_test_db=$1;
+    local recreate_db=$1;
+    local create_test_db=$2;
 
     if [ $create_test_db -eq 1 ]; then
         local test_db_name=`random_string 24`;
@@ -113,6 +114,10 @@ function test_get_or_create_db {
     else
         # name of test database expected to be defined in ODOO_TEST_CONF_FILE
         local test_db_name="$(odoo_get_conf_val db_name $ODOO_TEST_CONF_FILE)";
+    fi
+
+    if [ $recreate_db -eq 1 ] && odoo_db_exists $test_db_name; then
+        odoo_db_drop $test_db_name $ODOO_TEST_CONF_FILE 1>&2;
     fi
     echo "$test_db_name";
 }
@@ -199,20 +204,21 @@ function test_run_tests_handle_sigint {
 
 
 # Run tests
-# test_run_tests <create_test_db 1|0> <fail_on_warn 1|0> <with_coverage 1|0> <modules>
+# test_run_tests <recreate_db 1|0> <create_test_db 1|0> <fail_on_warn 1|0> <with_coverage 1|0> <modules>
 function test_run_tests {
-    local create_test_db=$1;
-    local fail_on_warn=$2;
-    local with_coverage=$3;
-    shift; shift; shift;
+    local recreate_db=$1;
+    local create_test_db=$2;
+    local fail_on_warn=$3;
+    local with_coverage=$4;
+    shift; shift; shift; shift;
 
     # Create new test database if required
-    local test_db_name="$(test_get_or_create_db $create_test_db)";
+    local test_db_name="$(test_get_or_create_db $recreate_db $create_test_db)";
     local test_log_file="${LOG_DIR:-.}/odoo.test.db.$test_db_name.log";
 
     # Remove log file if it is present before test, otherwise
     # it will be appended, wich could lead to incorrect test results
-    if [ -e $test_log_file ]; then
+    if [ ! -z "$test_log_file" ] && [ -e "$test_log_file" ]; then
         rm $test_log_file;
     fi
 
@@ -310,6 +316,7 @@ function test_run_pylint {
 # test_module [--tmp-dirs] [--create-test-db] -d <dir with addons to test>
 function test_module {
     local create_test_db=0;
+    local recreate_db=0;
     local fail_on_warn=0;
     local with_coverage=0;
     local with_coverage_report_html=;
@@ -327,6 +334,7 @@ function test_module {
 
     Options:
         --create-test-db    - Creates temporary database to run tests in
+        --recreate-db       - Recreate test database if it already exists
         --fail-on-warn      - if this option passed, then tests will fail even on warnings
         --tmp-dirs          - use temporary dirs for test related downloads and addons
         --no-rm-tmp-dirs    - not remove temporary directories that was created for this test
@@ -354,6 +362,9 @@ function test_module {
             ;;
             --create-test-db)
                 create_test_db=1;
+            ;;
+            --recreate-db)
+                recreate_db=1;
             ;;
             --fail-on-warn)
                 fail_on_warn=1;
@@ -406,8 +417,8 @@ function test_module {
     fi
 
     # Run tests
-    if test_run_tests ${create_test_db:-0} ${fail_on_warn:-0} \
-            ${with_coverage:-0} $modules;
+    if test_run_tests ${recreate_db:-0} ${create_test_db:-0} \
+        ${fail_on_warn:-0} ${with_coverage:-0} $modules;
     then
         local res=$?;
     else
@@ -419,7 +430,7 @@ function test_module {
     fi
 
     if [ ! -z "$with_coverage_report" ]; then
-        execv coverage report;
+        execv coverage report;  # --skip-covered;
     fi
     # ---------
